Move drawRCS methods into RCSController
