+++ SpacePhysics/SpacePhysics/Player/Ship.cs
@@ -13,7 +13,6 @@ namespace SpacePhysics.Player;
 public class Ship : CustomGameComponent
 {
   private AnimatedSprite thrustSprite;
-  private AnimatedSprite rcsSprite;
   private Texture2D thrustOverlay;
   private Texture2D thrustLensFlare;
 
@@ -45,6 +44,8 @@ public class Ship : CustomGameComponent
   public Ship(Func<float> opacity, bool allowInput, Alignment alignment, int layerIndex) : base(allowInput, alignment, layerIndex)
   {
     this.opacity = opacity;
+
+    components.Add(new RCSController(opacity));
   }
 
   public override void Initialize()
@@ -57,6 +58,8 @@ public class Ship : CustomGameComponent
     maxThrust = 600000f;
     engineEfficiency = 0.00000001f;
     rcsLerpSpeed = 30f;
+
+    base.Initialize();
   }
 
   public override void Load(ContentManager contentManager)
@@ -74,12 +77,7 @@ public class Ship : CustomGameComponent
       1f / 15f
     );
 
-    rcsSprite = new AnimatedSprite(
-      contentManager.Load<Texture2D>("Player/rcs-sheet"),
-      4,
-      1,
-      1f / 24f
-    );
+    base.Load(contentManager);
   }
 
   public override void Update()
@@ -87,7 +85,6 @@ public class Ship : CustomGameComponent
     if (state != State.Pause)
     {
       thrustSprite.Animate();
-      rcsSprite.Animate();
 
       Physics();
       Throttle();
@@ -97,9 +94,9 @@ public class Ship : CustomGameComponent
       RCS();
 
       electricity += deltaTime * 0.5f;
-    }
 
-    base.Update();
+      base.Update();
+    }
   }
 
   public override void Draw(SpriteBatch spriteBatch)
@@ -130,31 +127,7 @@ public class Ship : CustomGameComponent
 
     DrawThrust(spriteBatch);
 
-    // [0]: rotate left
-    // [1]: rotate right
-    // [2]: up
-    // [3]: down
-    // [4]: left
-    // [5]: right
-
-    DrawRCS(spriteBatch, new Vector2(30, -30), (float)Math.PI * 0.5f, rcsAmount[1] + rcsAmount[4]); // Bottom right
-    DrawRCS(spriteBatch, new Vector2(-33, -30), (float)-Math.PI * 0.5f, rcsAmount[0] + rcsAmount[5]); // Bottom left
-
-    DrawRCS(spriteBatch, new Vector2(-47, -30), (float)Math.PI * 0.5f, rcsAmount[0] + rcsAmount[4]); // Top right
-    DrawRCS(spriteBatch, new Vector2(44, -30), (float)-Math.PI * 0.5f, rcsAmount[1] + rcsAmount[5]); // Top left
-
-    // Docking mode
-    DrawRCS(spriteBatch, new Vector2(28, -32), (float)Math.PI, rcsAmount[2]); // Bottom left
-    DrawRCS(spriteBatch, new Vector2(-31, -32), (float)Math.PI, rcsAmount[2]); // Bottom right
-
-    DrawRCS(spriteBatch, new Vector2(28, 44), (float)Math.PI, rcsAmount[2]); // Top left
-    DrawRCS(spriteBatch, new Vector2(-31, 44), (float)Math.PI, rcsAmount[2]); // Top right
-
-    DrawRCS(spriteBatch, new Vector2(28, -47), 0f, rcsAmount[3]); // Top right retro
-    DrawRCS(spriteBatch, new Vector2(-32, -47), 0f, rcsAmount[3]); // Top left retro
-
-    DrawRCS(spriteBatch, new Vector2(28, 32), 0f, rcsAmount[3]); // Bottom right retro
-    DrawRCS(spriteBatch, new Vector2(-32, 32), 0f, rcsAmount[3]); // Bottom left retro
+    DrawAllRCS(spriteBatch);
 
     float thrustLensFlareScale = scale * thrustAmount;
 
@@ -287,37 +260,6 @@ public class Ship : CustomGameComponent
       thrustSprite.texture,
       adjustedPosition,
       thrustSprite.SourceRectangle,
-      Color.White * opacity(),
-      rotation,
-      origin,
-      thrustScale,
-      SpriteEffects.None,
-      0f
-    );
-  }
-
-  private void DrawRCS(SpriteBatch spriteBatch, Vector2 offsetOverride, float rotationOverride, float scaleOverride)
-  {
-    scaleOverride = Math.Clamp(scaleOverride, 0f, 1f);
-
-    float thrustScale = scaleOverride * scale * 0.2f;
-
-    Vector2 offset = new Vector2(0, scale) + offsetOverride;
-    Vector2 origin = new Vector2(rcsSprite.texture.Width / 2, rcsSprite.texture.Width);
-
-    float rotation = direction + rotationOverride;
-
-    Vector2 rotatedOffset = new Vector2(
-      offset.X * (float)Math.Cos(rotation) - offset.Y * (float)Math.Sin(rotation),
-      offset.X * (float)Math.Sin(rotation) + offset.Y * (float)Math.Cos(rotation)
-    );
-
-    Vector2 adjustedPosition = GameState.position + rotatedOffset;
-
-    spriteBatch.Draw(
-      rcsSprite.texture,
-      adjustedPosition,
-      rcsSprite.SourceRectangle,
       Color.White * opacity(),
       rotation,
       origin,
